requirements/base.txt
APScheduler==3.0.5
blinker==1.4
Flask==0.12.0
Flask-Admin==1.4.0
Flask-JWT-Extended==1.0.3
Flask-Limiter==0.8.5
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
itsdangerous==0.24
Jinja2==2.8
limits==1.0.9
MarkupSafe==0.23
oauthlib==1.0.3
psycopg2==2.6.1
PyJWT==1.4.0
python-memcached==1.57
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.5.0
simplejson==3.8.1
six==1.10.0
SQLAlchemy==1.0.11
twython==3.3.0
tzlocal==1.2
urllib3==1.12
Werkzeug==0.11.4
wheel==0.29.0
WTForms==2.1